Test: 期限切れのお知らせに対する既読登録のテストを追加し、既存のテストを整理
